refactor: Schedule 관련 레포지토리의 위치 수정
