+++ src/main/java/com/project/trainingdiary/repository/schedule/ScheduleRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.project.trainingdiary.repository;
+package com.project.trainingdiary.repository.schedule;
 
 import com.project.trainingdiary.dto.response.ScheduleResponseDto;
 import java.time.LocalDateTime;
