+++ src/main/java/com/project/trainingdiary/service/ScheduleService.java
@@ -18,9 +18,9 @@ import com.project.trainingdiary.exception.impl.ScheduleStatusNotOpenException;
 import com.project.trainingdiary.exception.impl.UserNotFoundException;
 import com.project.trainingdiary.model.ScheduleStatus;
 import com.project.trainingdiary.repository.PtContractRepository;
-import com.project.trainingdiary.repository.ScheduleRepository;
 import com.project.trainingdiary.repository.TraineeRepository;
 import com.project.trainingdiary.repository.TrainerRepository;
+import com.project.trainingdiary.repository.schedule.ScheduleRepository;
 import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
